Delete the route to display all statuses
It does not have any sort of pagination so it tries to fetch all
statuses from Redis, which is potentially a very expensive operation
depending on the number of elements.

In avant-basic this feature is completely broken because the number of
stored statuses is very large so there's no way Redis has time to get
them all in the 30 seconds that our app has to respond by default.
Because of this, there's no harm in removing this for us right now and
there's a chance it helps alleviating our apparently random Redis
issues.